Expose python test runner and test utils #11335
Previously, this relied on templating at the project level.
